[FJ-64] rename api to views
